Adding user crud

+++ app/Http/Controllers/BooksController.php
@@ -1,9 +1,7 @@
 <?php
 
 namespace App\Http\Controllers;
-
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Auth;
 use App\Models\Book;
 
 
@@ -13,14 +11,17 @@ class BooksController extends Controller
 
     public function __construct(){
         $this->middleware('auth');
+        $this->user = $user;
     }
 
 
     public function create(){
-        return view('books.create');  
+        $this->authorize('is_admin');
+        return view('books.create');
     }
 
     public function store(Request $request){
+        $this->authorize('is_admin');
         Book::create([
             'title' => $request->title,
             'author' => $request->author,
@@ -29,29 +30,38 @@ class BooksController extends Controller
             'category' => $request->category,
             'description' => $request->description,
             'publishing_company' => $request->publishing_company,
-            'amount' => $request->amount,
             'published_at' => $request->published_at,
         ]);
         return redirect('/show/books');
     }
 
-    public function show(){
+    public function show($id){
+        $this->authorize('is_admin');
+        $book = Book::findOrFail($id);
+        return view('books.show', ['book' => $book]);
+    }
+
+    public function showAll(){
+        $this->authorize('is_admin');
         $books = Book::all();
         return view('books.showAll',['books' => $books]);
     }
 
     public function destroy($id){
+        $this->authorize('is_admin');
         $book=Book::findOrFail($id);
         $book->delete();
         return redirect('/show/books');
     }
 
     public function edit($id){
+        $this->authorize('is_admin');
         $book = Book::findOrFail($id);
         return view('books.edit', ['book' => $book]);
     }
 
     public function update(Request $request, $id){
+        $this->authorize('is_admin');
         $book = Book::findOrFail($id);
         $book->update([
             'title' => $request->title,
@@ -61,7 +71,6 @@ class BooksController extends Controller
             'category' => $request->category,
             'description' => $request->description,
             'publishing_company' => $request->publishing_company,
-            'amount' => $request->amount,
             'published_at' => $request->published_at,
         ]);
         return redirect('/show/books');
